+++ CHANGES.md
@@ -1,3 +1,11 @@
+## 0.1.1 ##
+
+* Bug fixes
+* API changes (login and restore functions return void instead of boolean and
+  throw an error on login and restore failure)
+* Start using async/await
+* Remove promise shim and add promise runtime requirement to readme
+
 ## 0.1.0 ##
 
 * Initial public release